// src/gen/models/fSResponseDTO.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * STORE ME API
 * Hệ thống lưu trữ file cho phép người dùng upload, download và quản lý file của mình.

 * OpenAPI spec version: 1.0 beta
 */
import type { FSResponseDTOType } from './fSResponseDTOType';
import type { FileMetadataResponseDTO } from './fileMetadataResponseDTO';

export interface FSResponseDTO {
  id?: number;
  version?: number;
  createdAt?: string;
  updatedAt?: string;
  createdBy?: number;
  updatedBy?: number;
  name?: string;
  userId?: number;
  size?: number;
  ancestor?: number[];
  parentId?: number;
  lastAccessed?: string;
  deletedAt?: string;
  type?: FSResponseDTOType;
  fileMetadata?: FileMetadataResponseDTO;
  locked?: boolean;
  hidden?: boolean;
}