// src/gen/models/roleResponseDTO.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * STORE ME API
 * Hệ thống lưu trữ file cho phép người dùng upload, download và quản lý file của mình.

 * OpenAPI spec version: 1.0 beta
 */

export interface RoleResponseDTO {
  id?: number;
  version?: number;
  createdAt?: string;
  updatedAt?: string;
  createdBy?: number;
  updatedBy?: number;
  name?: string;
  description?: string;
}